keda-gcp-poc/consumer/requirements.txt
google-cloud-pubsub==2.27.1
